Klasse Admin in ReportEditScreen und ReportListTable aufgespalten
